connection: Add Transactor interface
Turns out that not everything wants a Protocol and a Transport...
some things just speak Packets. Probably this interface should be in a
separate file, and Connection should be called something a bit different
to denote that it's a Connection made of a Transport and a Protocol.

Also, interfaces should be named after their Method-er